+++ tests/test_polars_utils_schemas.py
@@ -69,13 +69,6 @@ class CleaningUtilsSchemas:
         ]
     )
 
-    reduce_dataset_to_earliest_file_per_month_schema = pl.Schema(
-        [
-            (CQCLClean.location_id, pl.String()),
-            (CQCLClean.cqc_location_import_date, pl.Date()),
-        ]
-    )
-
     create_banded_bed_count_column_schema = pl.Schema(
         [
             (IndCQC.location_id, pl.String()),